+++ src/component/Slider/Slider.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import { InputWrapperBaseProps, SliderProps as MantineSliderProps, Slider as MantineSlider, Input } from "@mantine/core";
+import { InputWrapperBaseProps, SliderProps as MantineSliderProps, Slider as MantineSlider, Input, GroupProps } from "@mantine/core";
+import { mergeSx } from "../../util";
 
 /**
  * Options for the {@link Slider} component.
@@ -18,16 +19,25 @@ export interface SliderProps extends MantineSliderProps {
     /**
      * Input description, displayed after label.
      */
-    "description"?: InputWrapperBaseProps["description"]
+    "description"?: InputWrapperBaseProps["description"],
+
+    /**
+     * Defines flex-grow property for each element, true -> 1, false -> 0
+     */
+    "grow"?: GroupProps["grow"]
 }
 
 /**
  * Mantine UI's {@link https://mantine.dev/core/slider Slider} wrapped in an {@link https://mantine.dev/core/input-wrapper InputWrapper}.
  */
-export function Slider({ label, hoverLabel, description, ...other }: SliderProps) {
+export function Slider({ label, hoverLabel, description, grow = false, sx, ...other }: SliderProps) {
     return (
-        <Input.Wrapper label={label} description={description}>
-            <MantineSlider label={hoverLabel} {...other} />
+        <Input.Wrapper label={label} description={description} sx={{
+            "flexGrow": grow ? 1 : 0
+        }}>
+            <MantineSlider label={hoverLabel} sx={mergeSx({
+                "flexGrow": grow ? 1 : 0
+            }, sx)} {...other} />
         </Input.Wrapper>
     );
 }